+++ wpull/hook.py
@@ -6,72 +6,9 @@ import sys
 
 from wpull.backport.logging import BraceMessage as __
 from wpull.database import Status
-import wpull.string
 
 
 _logger = logging.getLogger(__name__)
-
-
-def load_lua():
-    '''Load the Lua module.
-
-    .. seealso:: http://stackoverflow.com/a/8403467/1524507
-    '''
-    import DLFCN
-    sys.setdlopenflags(DLFCN.RTLD_NOW | DLFCN.RTLD_GLOBAL)
-    import lua
-    return lua
-
-
-try:
-    lua = load_lua()
-except (ImportError, AttributeError):
-    lua = None
-
-
-def to_lua_type(instance):
-    '''Convert instance to appropriate Python types for Lua.'''
-    return to_lua_table(to_lua_string(to_lua_number(instance)))
-
-
-def to_lua_string(instance):
-    '''If Lua, convert to bytes.'''
-    if sys.version_info[0] == 2:
-        return wpull.string.to_bytes(instance)
-    else:
-        return instance
-
-
-def to_lua_number(instance):
-    '''If Lua and Python 2, convert to long.'''
-    if sys.version_info[0] == 2:
-        if instance is True or instance is False:
-            return instance
-        elif isinstance(instance, int):
-            return long(instance)
-        elif isinstance(instance, list):
-            return list([to_lua_number(item) for item in instance])
-        elif isinstance(instance, tuple):
-            return tuple([to_lua_number(item) for item in instance])
-        elif isinstance(instance, dict):
-            return dict(
-                [(to_lua_number(key), to_lua_number(value))
-                    for key, value in instance.items()])
-        return instance
-    else:
-        return instance
-
-
-def to_lua_table(instance):
-    '''If Lua and instance is ``dict``, convert to Lua table.'''
-    if isinstance(instance, dict):
-        table = lua.eval('{}')
-
-        for key, value in instance.items():
-            table[to_lua_table(key)] = to_lua_table(value)
-
-        return table
-    return instance
 
 
 class HookDisconnected(Exception):
@@ -147,44 +84,6 @@ class LegacyCallbacks(object):
        scripts; the arugment signature will be adjusted automatically.
     '''
 
-    @staticmethod
-    def handle_response1(url_info, http_info):
-        '''Return an action to handle the response.
-
-        Args:
-            url_info (dict): A mapping containing the same information in
-                :class:`.url.URLInfo`.
-            http_info (dict): A mapping containing the same information
-                in :class:`.http.request.Response`.
-
-        .. deprecated:: Scripting-API-2
-
-        Returns:
-            str: A value from :class:`Actions`. The default is
-            :attr:`Actions.NORMAL`.
-        '''
-        return Actions.NORMAL
-
-    @staticmethod
-    def handle_error1(url_info, error_info):
-        '''Return an action to handle the error.
-
-        Args:
-            url_info (dict): A mapping containing the same information in
-                :class:`.url.URLInfo`.
-            http_info (dict): A mapping containing the keys:
-
-                * ``error``: The name of the exception (for example,
-                  ``ProtocolError``)
-
-        .. deprecated:: Scripting-API-2
-
-        Returns:
-            str: A value from :class:`Actions`. The default is
-            :attr:`Actions.NORMAL`.
-        '''
-        return Actions.NORMAL
-
 
 class Callbacks(LegacyCallbacks):
     '''Callback hooking instance.
@@ -195,10 +94,10 @@ class Callbacks(LegacyCallbacks):
             Default: 2.
 
     '''
-    AVAILABLE_VERSIONS = to_lua_number((1, 2,))
+    AVAILABLE_VERSIONS = (2,)
 
     def __init__(self):
-        self._version = to_lua_number(2)
+        self._version = 2
 
     @property
     def version(self):
@@ -303,17 +202,6 @@ class Callbacks(LegacyCallbacks):
         '''
         return Actions.NORMAL
 
-    original_handle_response = handle_response
-
-    def call_handle_response(self, url_info, record_info, http_info):
-        '''Call the correct :meth:`handle_response`.'''
-
-        if self.version >= 2 \
-           or self.original_handle_response == self.handle_response:
-            return self.handle_response(url_info, record_info, http_info)
-        else:
-            return self.handle_response(url_info, http_info)
-
     @staticmethod
     def handle_error(url_info, record_info, error_info):
         '''Return an action to handle the error.
@@ -336,17 +224,6 @@ class Callbacks(LegacyCallbacks):
             :attr:`Actions.NORMAL`.
         '''
         return Actions.NORMAL
-
-    original_handle_error = handle_error
-
-    def call_handle_error(self, url_info, record_info, error_info):
-        '''Call the correct :meth:`handle_error`.'''
-
-        if self.version >= 2 \
-           or self.original_handle_error == self.handle_error:
-            return self.handle_error(url_info, record_info, error_info)
-        else:
-            return self.handle_error(url_info, error_info)
 
     @staticmethod
     def get_urls(filename, url_info, document_info):
@@ -431,48 +308,10 @@ class HookEnvironment(object):
         actions (:class:`Actions`): The Actions instance.
         callbacks (:class:`Callbacks`): The Callback instance.
     '''
-    def __init__(self, factory, is_lua=False):
+    def __init__(self, factory):
         self.factory = factory
-        self.is_lua = is_lua
         self.actions = Actions()
         self.callbacks = Callbacks()
-
-    def to_script_native_type(self, instance):
-        '''Convert the instance recursively to native script types.
-
-        If the script is Lua, call :func:`to_lua_type`. Otherwise,
-        returns instance unchanged.
-
-        Returns:
-            instance
-        '''
-        if self.is_lua:
-            return to_lua_type(instance)
-        return instance
-
-    def get_from_native_dict(self, instance, key, default=None):
-        '''Try to get from the mapping a value.
-
-        This method will try to determine whether a Lua table or
-        ``dict`` is given.
-        '''
-        try:
-            instance.attribute_should_not_exist
-        except AttributeError:
-            return instance.get(key, default)
-        else:
-            # Check if key exists in Lua table
-            value_1 = instance[self.to_script_native_type(key)]
-
-            if value_1 is not None:
-                return value_1
-
-            value_2 = getattr(instance, self.to_script_native_type(key))
-
-            if value_1 is None and value_2 is None:
-                return default
-            else:
-                return value_1
 
     def connect_hooks(self):
         '''Connect callbacks to hooks.'''
@@ -506,7 +345,7 @@ class HookEnvironment(object):
             self._scrape_document)
 
     def _resolve_dns(self, host, port):
-        answer = self.callbacks.resolve_dns(to_lua_type(host))
+        answer = self.callbacks.resolve_dns(host)
 
         _logger.debug(__('Resolve hook returned {0}', answer))
 
@@ -519,47 +358,40 @@ class HookEnvironment(object):
         self.callbacks.engine_run()
 
     def _exit_status(self, exit_status):
-        return self.callbacks.exit_status(
-            self.to_script_native_type(exit_status))
+        return self.callbacks.exit_status(exit_status)
 
     def _finishing_statistics(self, start_time, stop_time, files, size):
         self.callbacks.finishing_statistics(
-            to_lua_type(start_time),
-            to_lua_type(stop_time),
-            to_lua_type(files),
-            to_lua_type(size)
+            start_time,
+            stop_time,
+            files,
+            size
         )
 
     def _wait_time(self, seconds):
-        if self.is_lua:
-            seconds = to_lua_type(seconds)
         return self.callbacks.wait_time(seconds)
 
     def _queued_url(self, url_info):
-        url_info_dict = self.to_script_native_type(url_info.to_dict())
+        url_info_dict = url_info.to_dict()
 
         self.callbacks.queued_url(url_info_dict)
 
     def _dequeued_url(self, url_info, url_record):
-        url_info_dict = self.to_script_native_type(url_info.to_dict())
-
+        url_info_dict = url_info.to_dict()
         record_info_dict = url_record.to_dict()
-        record_info_dict = self.to_script_native_type(record_info_dict)
 
         self.callbacks.dequeued_url(url_info_dict, record_info_dict)
 
     def _should_fetch(self, url_info, url_record, verdict, reason_slug,
                       test_info):
-        url_info_dict = self.to_script_native_type(url_info.to_dict())
+        url_info_dict = url_info.to_dict()
 
         record_info_dict = url_record.to_dict()
-        record_info_dict = self.to_script_native_type(record_info_dict)
 
         reasons = {
             'filters': test_info['map'],
             'reason': reason_slug,
         }
-        reasons = self.to_script_native_type(reasons)
 
         verdict = self.callbacks.accept_url(
             url_info_dict, record_info_dict, verdict, reasons)
@@ -569,14 +401,11 @@ class HookEnvironment(object):
         return verdict
 
     def _handle_response(self, request, response, url_item):
-        url_info_dict = self.to_script_native_type(
-            request.url_info.to_dict()
-        )
-        url_record_dict = self.to_script_native_type(
-            url_item.url_record.to_dict()
-        )
-        response_info_dict = self.to_script_native_type(response.to_dict())
-        action = self.callbacks.call_handle_response(
+        url_info_dict = request.url_info.to_dict()
+        url_record_dict = url_item.url_record.to_dict()
+
+        response_info_dict = response.to_dict()
+        action = self.callbacks.handle_response(
             url_info_dict, url_record_dict, response_info_dict
         )
 
@@ -595,16 +424,12 @@ class HookEnvironment(object):
             raise NotImplementedError()
 
     def _handle_error(self, request, url_item, error):
-        url_info_dict = self.to_script_native_type(
-            request.url_info.to_dict()
-        )
-        url_record_dict = self.to_script_native_type(
-            url_item.url_record.to_dict()
-        )
-        error_info_dict = self.to_script_native_type({
+        url_info_dict = request.url_info.to_dict()
+        url_record_dict = url_item.url_record.to_dict()
+        error_info_dict = {
             'error': error.__class__.__name__,
-        })
-        action = self.callbacks.call_handle_error(
+        }
+        action = self.callbacks.handle_error(
             url_info_dict, url_record_dict, error_info_dict
         )
 
@@ -623,10 +448,9 @@ class HookEnvironment(object):
             raise NotImplementedError()
 
     def _scrape_document(self, request, response, url_item):
-        to_native = self.to_script_native_type
-        url_info_dict = to_native(request.url_info.to_dict())
-        document_info_dict = to_native(response.body.to_dict())
-        filename = to_native(response.body.name)
+        url_info_dict = request.url_info.to_dict()
+        document_info_dict = response.body.to_dict()
+        filename = response.body.name
 
         new_url_dicts = self.callbacks.get_urls(
             filename, url_info_dict, document_info_dict)
@@ -636,29 +460,16 @@ class HookEnvironment(object):
         if not new_url_dicts:
             return
 
-        if to_native(1) in new_url_dicts:
-            # Lua doesn't have sequences
-            for i in itertools.count(1):
-                new_url_dict = new_url_dicts[to_native(i)]
-
-                _logger.debug(__('Got lua new url info {0}', new_url_dict))
-
-                if new_url_dict is None:
-                    break
-
-                self._add_hooked_url(url_item, new_url_dict)
-        else:
-            for new_url_dict in new_url_dicts:
-                self._add_hooked_url(url_item, new_url_dict)
+        for new_url_dict in new_url_dicts:
+            self._add_hooked_url(url_item, new_url_dict)
 
     def _add_hooked_url(self, url_item, new_url_dict):
         '''Process the ``dict`` from the script and add the URLs.'''
-        to_native = self.to_script_native_type
-        url = new_url_dict[to_native('url')]
-        link_type = self.get_from_native_dict(new_url_dict, 'link_type')
-        inline = self.get_from_native_dict(new_url_dict, 'inline')
-        post_data = self.get_from_native_dict(new_url_dict, 'post_data')
-        replace = self.get_from_native_dict(new_url_dict, 'replace')
+        url = new_url_dict['url']
+        link_type = new_url_dict.get('link_type')
+        inline = new_url_dict.get('inline')
+        post_data = new_url_dict.get('post_data')
+        replace = new_url_dict.get('replace')
 
         assert url